Restructure into core lib, create web module with an embedded fs
