+++ .gitignore
@@ -1,0 +1,3 @@
+embc
+embd
+outputs.dat
