+++ app/index.ts
@@ -1,0 +1,2 @@
+// App will live in this directory and get built into web/dist for use in the
+// embedded filesystem.
